+++ README.md
@@ -1,4 +1,4 @@
-# NoxUtils
+# NoxesiumAPI
 This is an API with support for [Skript](https://github.com/SkriptLang/Skript) in mind. You can always ask me to add something or check out [NoxesiumUtils by SuperNeon4ik](https://github.com/SuperNeon4ik/NoxesiumUtils/tree/master)
 
 ## Features
@@ -6,8 +6,12 @@ This is an API with support for [Skript](https://github.com/SkriptLang/Skript) i
 - Qib system
 - Client-side glow for entities and blocks
 
+## Requirements
+- Paper 1.21+ (or a fork)
+- CommandAPI
+
 ## Planned updates
-- [Update to Noxesium 2.4.0/1.21.2](https://github.com/BlockybombDEV/NoxUtils/pull/1)
+- Nothing. Yay!
 
 ## Credits
 - [Noxesium by Noxcrew](https://github.com/Noxcrew/noxesium) - Noxesium lol
