+++ src/main/kotlin/me/iris/noxesiumapi/serverrules/ResetRules.kt
@@ -1,115 +1,112 @@
-package me.iris.noxUtils.serverrules
+package me.iris.noxesiumapi.serverrules
 
 import com.noxcrew.noxesium.api.protocol.rule.ServerRuleIndices
-import com.noxcrew.noxesium.paper.api.rule.GraphicsType
 import com.noxcrew.noxesium.paper.api.rule.RemoteServerRule
-import me.iris.noxUtils.NoxUtils
-import me.iris.noxUtils.NoxUtils.Companion.customCreativeItems
+import me.iris.noxesiumapi.NoxesiumAPI
 import org.bukkit.entity.Player
-import org.bukkit.inventory.ItemStack
 
 /**
- * Set the value of a server rule for a [Player]
+ * Reset the value of a server rule for a [Player]
  */
-public open class SetRules(private val player: Player) {
+public class ResetRules(private val player: Player) {
 
-    private var manager = NoxUtils.noxesiumManager
+    private var manager = NoxesiumAPI.noxesiumManager
 
-    public fun disableSpinAttackCollision(value: Boolean) {
+    public fun disableSpinAttackCollision() {
         val rule: RemoteServerRule<Any>? = manager.getServerRule(player, ServerRuleIndices.DISABLE_SPIN_ATTACK_COLLISIONS)
 
-        rule!!.setValue(value)
+        rule!!.reset()
         manager.updateServerRules(player)
     }
 
-    public fun heldItemNameOffset(value: Int) {
+    public fun heldItemNameOffset() {
         val rule: RemoteServerRule<Any>? = manager.getServerRule(player, ServerRuleIndices.HELD_ITEM_NAME_OFFSET)
 
-        rule!!.setValue(value)
+        rule!!.reset()
         manager.updateServerRules(player)
     }
 
-    public fun cameraLocked(value: Boolean) {
+    public fun cameraLocked() {
         val rule: RemoteServerRule<Any>? = manager.getServerRule(player, ServerRuleIndices.CAMERA_LOCKED)
 
-        rule!!.setValue(value)
+        rule!!.reset()
         manager.updateServerRules(player)
     }
 
-    public fun disableVanillaMusic(value: Boolean) {
+    public fun disableVanillaMusic() {
         val rule: RemoteServerRule<Any>? = manager.getServerRule(player, ServerRuleIndices.DISABLE_VANILLA_MUSIC)
 
-        rule!!.setValue(value)
+        rule!!.reset()
         manager.updateServerRules(player)
     }
 
-    public fun disableBoatCollisions(value: Boolean) {
+    public fun disableBoatCollisions() {
         val rule: RemoteServerRule<Any>? = manager.getServerRule(player, ServerRuleIndices.DISABLE_BOAT_COLLISIONS)
 
-        rule!!.setValue(value)
+        rule!!.reset()
         manager.updateServerRules(player)
     }
 
-    public fun handItemOverride(value: ItemStack) {
+    public fun handItemOverride() {
         val rule: RemoteServerRule<Any>? = manager.getServerRule(player, ServerRuleIndices.HAND_ITEM_OVERRIDE)
 
-        rule!!.setValue(value)
+        rule!!.reset()
         manager.updateServerRules(player)
     }
 
-    public fun disableUiOptimizations(value: Boolean) {
+    public fun disableUiOptimizations() {
         val rule: RemoteServerRule<Any>? = manager.getServerRule(player, ServerRuleIndices.DISABLE_UI_OPTIMIZATIONS)
 
-        rule!!.setValue(value)
+        rule!!.reset()
         manager.updateServerRules(player)
     }
 
-    public fun showMapInUi(value: Boolean) {
+    public fun showMapInUi() {
         val rule: RemoteServerRule<Any>? = manager.getServerRule(player, ServerRuleIndices.SHOW_MAP_IN_UI)
 
-        rule!!.setValue(value)
+        rule!!.reset()
         manager.updateServerRules(player)
     }
 
-    public fun disableDeferredChunkUpdates(value: Boolean) {
+    public fun disableDeferredChunkUpdates() {
         val rule: RemoteServerRule<Any>? = manager.getServerRule(player, ServerRuleIndices.DISABLE_DEFERRED_CHUNK_UPDATES)
 
-        rule!!.setValue(value)
+        rule!!.reset()
         manager.updateServerRules(player)
     }
 
     public fun customCreativeItems() {
         val rule: RemoteServerRule<Any>? = manager.getServerRule(player, ServerRuleIndices.CUSTOM_CREATIVE_ITEMS)
 
-        rule!!.setValue(customCreativeItems)
+        rule!!.reset()
         manager.updateServerRules(player)
     }
 
-    public fun overrideGraphicsMode(value: GraphicsType) {
+    public fun overrideGraphicsMode() {
         val rule: RemoteServerRule<Any>? = manager.getServerRule(player, ServerRuleIndices.OVERRIDE_GRAPHICS_MODE)
 
-        rule!!.setValue(value)
+        rule!!.reset()
         manager.updateServerRules(player)
     }
 
-    public fun enableSmootherClientTrident(value: Boolean) {
+    public fun enableSmootherClientTrident() {
         val rule: RemoteServerRule<Any>? = manager.getServerRule(player, ServerRuleIndices.ENABLE_SMOOTHER_CLIENT_TRIDENT)
 
-        rule!!.setValue(value)
+        rule!!.reset()
         manager.updateServerRules(player)
     }
 
-    public fun disableMapInUi(value: Boolean) {
+    public fun disableMapInUi() {
         val rule: RemoteServerRule<Any>? = manager.getServerRule(player, ServerRuleIndices.DISABLE_MAP_UI)
 
-        rule!!.setValue(value)
+        rule!!.reset()
         manager.updateServerRules(player)
     }
 
-    public fun riptideCoyoteTime(value: Int) {
+    public fun riptideCoyoteTime() {
         val rule: RemoteServerRule<Any>? = manager.getServerRule(player, ServerRuleIndices.RIPTIDE_COYOTE_TIME)
 
-        rule!!.setValue(value)
+        rule!!.reset()
         manager.updateServerRules(player)
     }
